Moving model data into src/
This may change in future if it's going to be generated from datasets.
